+++ .gitignore
@@ -20,3 +20,4 @@ dist
 *.log
 *.local
 .cache
+tsconfig.tsbuildinfo
