+++ apps/admin/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@repo/utils": "*",
+    "@repo/backend-types": "*",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.12",
