+++ apps/backend/package.json
@@ -31,6 +31,7 @@
   "homepage": "https://github.com/LukaszGrela/master-word#readme",
   "dependencies": {
     "@repo/utils": "*",
+    "@repo/backend-types": "*",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
