+++ apps/backend/src/db/schemas/UnknownWordSchema.ts
@@ -10,7 +10,7 @@ const UnknownWordSchema = new Schema({
       new Schema({
         language: { type: SupportedLanguage, required: true },
         word: { type: Types.String, required: true },
-        length: Types.Number,
+        length: { type: Types.Number, required: true },
       }),
     ],
   },
