+++ apps/backend/src/db/types.ts
@@ -5,15 +5,3 @@ export type TConfig = {
     db: string;
   };
 };
-
-export interface IDictionaryEntry {
-  language: string;
-  length: number;
-  letter: string;
-  words?: string[];
-}
-
-export interface IUnknownWordEntry {
-  date: Date;
-  words: { language: string; word: string; length?: number }[];
-}
