+++ apps/backend/src/types.ts
@@ -4,43 +4,3 @@ export const guardSupportedLanguages = (
 ): test is TSupportedLanguages => {
   return typeof test === 'string' && (test === 'pl' || test === 'en');
 };
-// the game
-// Wrong - X, Misplaced - M, Correct - C
-export type TValidationChar = 'X' | 'M' | 'C';
-export type TGameState = {
-  word: string[];
-  validated: TValidationChar[];
-};
-
-export type TGameSessionFinished = {
-  finished: true;
-  timestamp_finish: string;
-};
-export type TGameSessionIncomplete = {
-  finished: false;
-};
-
-export type TScore = {
-  score: number;
-  timeMs: number;
-  attempts: number;
-};
-
-export type TGameSession = {
-  language: TSupportedLanguages;
-
-  timestamp_start: string;
-  max_attempts: number;
-  attempt: number;
-  word: string;
-  word_length: number;
-  state: TGameState[];
-  guessed: boolean;
-  score: number;
-} & (TGameSessionFinished | TGameSessionIncomplete);
-
-export type TGameSessionRecord = {
-  session: string;
-  highest?: TScore;
-  game: TGameSession;
-};
