+++ apps/frontend/package.json
@@ -22,6 +22,8 @@
   },
   "devDependencies": {
     "@repo/config-typescript": "*",
+    "@repo/backend-types": "*",
+    "@repo/common-types": "*",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
