--- apps/frontend/src/types/common.ts
@@ -1,3 +1,0 @@
-export interface IDictionary<T = unknown> {
-  [key: string]: T;
-}
