+++ packages/backend-types/README.md
@@ -1,0 +1,3 @@
+# Shared Backend Types
+
+Package will contain backend types that are shared across all APPS.
