+++ packages/backend-types/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@repo/backend-types",
+  "version": "0.3.1",
+  "private": true,
+  "license": "MIT",
+  "exports": {
+    ".": "./src/index.ts",
+    "./db": "./src/db.ts",
+    "./dictionary": "./src/dictionary.ts",
+    "./enums": "./src/enums.ts"
+  },
+  "devDependencies": {
+    "@repo/config-typescript": "*",
+    "eslint": "^8.56.0",
+    "typescript": "^5.2.2"
+  }
+}
