+++ packages/backend-types/src/db.ts
@@ -1,0 +1,16 @@
+export interface IDictionaryEntry {
+  language: string;
+  length: number;
+  letter: string;
+  words?: string[];
+}
+
+export interface IUnknownWordWordsItem {
+  language: string;
+  word: string;
+  length: number;
+}
+export interface IUnknownWordEntry {
+  date: Date;
+  words: IUnknownWordWordsItem[];
+}
