+++ packages/common-types/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@repo/common-types",
+  "version": "0.0.2",
+  "private": true,
+  "license": "MIT",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "devDependencies": {
+    "@repo/config-typescript": "*",
+    "eslint": "^8.56.0",
+    "typescript": "^5.2.2"
+  }
+}
